Move detention requests from Judge routes
